refactor: change generated `cli_info` file path and update imports in app_command_runner.dart, checksum.dart, and grind.dart
