+++ build.yaml
@@ -14,4 +14,4 @@ targets:
             - duplicate_ignore
       cli_info_builder:
         options:
-          output: "lib/src/app/gen/cli_info.g.dart"
+          output: "lib/src/app/gen/cli_info.dart"
