+++ lib/src/app/app_command_runner.dart
@@ -7,7 +7,7 @@ import 'package:dvmx/src/app/commands/list_command.dart';
 import 'package:dvmx/src/app/commands/releases_command.dart';
 import 'package:dvmx/src/app/commands/uninstall_command.dart';
 import 'package:dvmx/src/app/commands/use_command.dart';
-import 'package:dvmx/src/app/gen/cli_info.g.dart';
+import 'package:dvmx/src/app/gen/cli_info.dart';
 import 'package:dvmx/src/app/models/exit_status.dart';
 import 'package:dvmx/src/app/servicies/console_service.dart';
 import 'package:mason_logger/mason_logger.dart';
