+++ lib/src/app/gen/cli_info.dart
